fix(vim): fix spurious 'g' character
increase 'ttimeoutlen' to avoid having the character under the cursor
(the first character in the file with the current settings) being
sometimes replaced by the character 'g'.

opening a file with 'vim -E <file>' reveals the issue:

:]11;rgb:1e1e/1e1e/2020

the key codes are sometimes delayed (e.g. on a remote server) and are
handled improperly. the sequence 'rg' results in the observed behaviour.